// src/models/tiles/table-links.scss.d.ts
// cf. https://mattferderer.com/use-sass-variables-in-typescript-and-javascript

export interface ITableLinksStyleExports {
  linkColor0Light: string;
  linkColor0Dark: string;
  linkColor1Light: string;
  linkColor1Dark: string;
  linkColor2Light: string;
  linkColor2Dark: string;
  linkColor3Light: string;
  linkColor3Dark: string;
  linkColor4Light: string;
  linkColor4Dark: string;
  linkColor5Light: string;
  linkColor5Dark: string;
}

export const styles: ITableLinksStyleExports;

export default styles;